+++ libs/__init__.py
@@ -1,11 +1,14 @@
 # pylint:disable=wrong-import-position
 """库文件"""
-__all__ = ["START", "OUTPUT", "get_user_info", "DrawBest", "load_config"]
+__all__ = [
+    "START", "OUTPUT", "get_user_info", "music_list_dispatcher", 
+    "DrawBest", "Config", "load_config"
+]
 
 import time
 START = time.time()
 
 from .consts import OUTPUT
-from .api import get_user_info
+from .api import get_user_info, music_list_dispatcher
 from .image import DrawBest
-from .config import load_config
+from .config import Config, load_config