+++ libs/api.py
@@ -115,7 +115,6 @@ class MusicList(List[Music]):
 
 async def get_music_list() -> MusicList:
     """获取所有数据"""
-    # MusicData
     print("[INFO]正在与服务器通信...(获取曲目信息)")
     try:
         try:
@@ -138,27 +137,27 @@ async def get_music_list() -> MusicList:
             '将内容保存为 "music_data.json" 存放在 "static" 目录下并重启此工具'
         )
         raise
-    # ChartStats
+
     print("[INFO]正在与服务器通信...(获取谱面信息)")
     try:
         try:
             chart_stats = await mai_api.chart_stats()
             await writefile(chart_file, chart_stats)
         except asyncio.exceptions.TimeoutError:
-            print('[ERROR]从diving-fish获取maimaiDX数据获取超时，正在使用yuzuapi中转获取单曲数据')
+            print('[ERROR]从diving-fish获取maimaiDX谱面数据超时，正在使用yuzuapi中转获取谱面数据')
             chart_stats = await mai_api.transfer_chart()
             await writefile(chart_file, chart_stats)
         except UnknownError:
-            print('[ERROR]从diving-fish获取maimaiDX单曲数据获取错误。已切换至本地暂存文件')
+            print('[ERROR]从diving-fish获取maimaiDX谱面数据错误，请检查网络环境。已切换至本地暂存文件')
             chart_stats = await openfile(chart_file)
         except Exception: # pylint: disable=broad-exception-caught
             print(f'[ERROR]Error: {traceback.format_exc()}')
-            print('maimaiDX数据获取错误，请检查网络环境。已切换至本地暂存文件')
+            print('maimaiDX谱面数据获取失败，请检查网络环境。已切换至本地暂存文件')
             chart_stats = await openfile(chart_file)
     except FileNotFoundError:
         print(
             '[FATAL]未找到文件，请自行使用浏览器访问 "https://www.diving-fish.com/api/maimaidxprober/chart_stats" '
-            '将内容保存为 "chart_stats.json" 存放在 "static" 目录下并重启此工具'
+            '将内容保存为 "music_chart.json" 存放在 "static" 目录下并重启此工具'
         )
         raise
 
@@ -172,6 +171,86 @@ async def get_music_list() -> MusicList:
         total_list.append(Music(stats=_stats, **music))
 
     return total_list
+
+async def get_local_music_list() -> MusicList:
+    """获取本地曲目数据"""
+    print("[INFO]正在读取本地曲目数据...")
+    try:
+        music_data = await openfile(music_file)
+    except FileNotFoundError:
+        print(
+            '[ERROR]未找到文件，即将向服务器请求曲目信息。\n'
+            '[INFO]正在与服务器通信...(获取曲目信息)'
+        )
+        try:
+            music_data = await mai_api.music_data()
+            await writefile(music_data, music_file)
+        except asyncio.exceptions.TimeoutError:
+            print('[ERROR]从diving-fish获取maimaiDX曲目数据超时，正在使用yuzuapi中转获取曲目数据')
+            music_data = await mai_api.transfer_music()
+            await writefile(music_data, music_file)
+        except UnknownError:
+            print(
+                '[FATAL]从diving-fish获取maimaiDX曲目数据错误。请检查网络环境。\n'
+                '[FATAL]请自行使用浏览器访问 "https://www.diving-fish.com/api/maimaidxprober/music_data"'
+                '将内容保存为 "music_data.json" 存放在 "static" 目录下并重启此工具'
+            )
+            raise
+        except Exception: # pylint: disable=broad-exception-caught
+            print(f'[FATAL]Error: {traceback.format_exc()}')
+            print(
+                '[FATAL]maimaiDX数据获取错误，请检查网络环境。\n'
+                '[FATAL]请自行使用浏览器访问 "https://www.diving-fish.com/api/maimaidxprober/music_data"'
+                '将内容保存为 "music_data.json" 存放在 "static" 目录下并重启此工具'
+            )
+            raise
+
+    print("[INFO]正在读取本地谱面数据...")
+    try:
+        chart_stats = await openfile(chart_file)
+    except FileNotFoundError:
+        print(
+            '[ERROR]未找到文件，即将向服务器请求谱面信息。\n'
+            '[INFO]正在与服务器通信...(获取谱面信息)'
+        )
+        try:
+            chart_stats = await mai_api.chart_stats()
+            await writefile(chart_file, chart_stats)
+        except asyncio.exceptions.TimeoutError:
+            print('[ERROR]从diving-fish获取maimaiDX谱面数据超时，正在使用yuzuapi中转获取谱面数据')
+            chart_stats = await mai_api.transfer_chart()
+            await writefile(chart_file, chart_stats)
+        except UnknownError:
+            print(
+                '[FATAL]从diving-fish获取maimaiDX谱面数据错误。请检查网络环境。\n'
+                '[FATAL]请自行使用浏览器访问 "https://www.diving-fish.com/api/maimaidxprober/chart_stats"'
+                '将内容保存为 "music_chart.json" 存放在 "static" 目录下并重启此工具'
+            )
+            raise
+        except Exception: # pylint: disable=broad-exception-caught
+            print(f'[FATAL]Error: {traceback.format_exc()}')
+            print(
+                '[FATAL]maimaiDX数据获取错误，请检查网络环境。\n'
+                '[FATAL]请自行使用浏览器访问 "https://www.diving-fish.com/api/maimaidxprober/chart_stats"'
+                '将内容保存为 "music_chart.json" 存放在 "static" 目录下并重启此工具'
+            )
+            raise
+
+    total_list: MusicList = MusicList()
+    for music in music_data:
+        if music['id'] in chart_stats['charts']:
+            _stats = [_data if _data else None for _data in chart_stats['charts'][music['id']]] \
+                if {} in chart_stats['charts'][music['id']] else chart_stats['charts'][music['id']]
+        else:
+            _stats = None
+        total_list.append(Music(stats=_stats, **music))
+
+    return total_list
+
+async def music_list_dispatcher(local_first: bool) -> MusicList:
+    """曲目列表"""
+    return await get_local_music_list() if local_first else await get_music_list()
+
 
 async def get_user_info(username: str) -> Optional[dict]:
     """获取用户信息"""